+++ docs/infra/editor/components/RicosEditorWithMobile.tsx
@@ -1,0 +1,4 @@
+import { RicosEditor } from 'ricos-editor';
+import { withMobileDetection } from './withMobileDetect';
+
+export const RicosEditorWithMobile = withMobileDetection(RicosEditor);